[Add] 유저 자신이 작성한 댓글 삭제

+++ src/main/java/jungle/fairyTeller/board/dto/CommentDto.java
@@ -12,7 +12,7 @@ import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CommentDto {
-    private Long commentId;
+    private Integer commentId;
     private Integer boardId;
     private String content;
     private String author;
